+++ Sources/BaseFoundationiOS/Templates/Enums/TokenType.swift
@@ -1,0 +1,14 @@
+//
+//  TokenType.swift
+//
+//
+//  Created by pablo borquez on 04/11/23.
+//
+
+import Foundation
+
+public enum TokenType: String {
+    case bearer = "Bearer"
+    case undefined
+
+}